refactor: 컴포넌트 폴더로 분리
