+++ src/components/board/MainBestItem.tsx
