+++ src/components/board/MainBoard.tsx
@@ -2,7 +2,7 @@ import { faChevronRight } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { useHistory } from 'react-router-dom';
 import styled from 'styled-components';
-import { MainBoardPost } from '../types/board';
+import { MainBoardPost } from '../../types/board';
 import MainBoardItem from './MainBoardItem';
 
 interface MainBoardProps {
