+++ src/components/board/MainBoardItem.tsx
