+++ src/routes/board/NormalBoard.tsx
@@ -3,10 +3,10 @@ import { useParams } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import { fetchPostList } from '../../api';
 import { userInfoState } from '../../atoms';
-import BoardHeader from '../../components/BoardHeader';
-import PostList from '../../components/PostList';
-import BoardListFooter from '../../components/BoardListFooter';
-import NavContainer from '../../components/NavContainer';
+import BoardHeader from '../../components/board/BoardHeader';
+import PostList from '../../components/board/PostList';
+import BoardListFooter from '../../components/board/BoardListFooter';
+import NavContainer from '../../components/board/NavContainer';
 import { BoardURLParams, Post } from '../../types/board';
 
 function NormalBoard() {
